Style mobile location dropdown.
