+++ client/components/navigation/header.jsx
@@ -5,16 +5,15 @@ import Tabs from 'material-ui/Tabs/Tabs';
 import Toolbar from 'material-ui/Toolbar';
 import ToolbarTitle from 'material-ui/Toolbar/ToolbarTitle';
 import ToolbarGroup from 'material-ui/Toolbar/ToolbarGroup';
-import SelectField from 'material-ui/SelectField';
-import MenuItem from 'material-ui/MenuItem';
 
-import MenuButton from './menu-button';
 import Responsive from '../common/responsive';
+import MenuButton from './menu-button';
+import LocationDropDown from './location-dropdown';
 
+import { styles } from './styles';
 import { MOBILE_WIDTH_BREAKPOINT } from '../common/styles';
 import { formatForDisplay } from '../../utils';
 import { applyStyles } from '../../config/composition';
-import { styles } from './styles';
 
 const Header = (props) => {
   const { params, location, locations, actions, siteNavOpen } = props;
@@ -30,23 +29,6 @@ const Header = (props) => {
       style={styles.toolbarTab}/>
   );
 
-  const renderLocationSelection = (renderedLocation, index) => (
-    <MenuItem
-      key={index}
-      value={index}
-      primaryText={formatForDisplay(renderedLocation)}
-      onClick={actions.emitLocationIndexUpdate}/>
-  );
-
-  const locationsDropDown = (
-    <SelectField
-      value={locations.indexOf(params.location)}
-      onChange={() => {}}
-      floatingLabelText='Select a Location'>
-        {locations.map(renderLocationSelection)}
-    </SelectField>
-  );
-
   return fullscreen === 'true' ? null : (
     <Toolbar style={styles.toolbar}>
       <ToolbarGroup firstChild={true}>
@@ -58,7 +40,7 @@ const Header = (props) => {
       <ToolbarGroup style={styles.toolbarTabs}>
         <Responsive
           mobileBreakpoint={MOBILE_WIDTH_BREAKPOINT}
-          mobileAlt={locationsDropDown}
+          mobileAlt={<LocationDropDown {...props}/>}
           {...props}>
             <Tabs value={locations.indexOf(params.location)}>
               {locations.map(renderLocationTab)}
